The initial version of script getsj.py
Background:
When access a website, it will download  relative css and js files. I
access some special sites so frequently, but it refreshes  sllowlly.
So I try to download the entire website and access it locally. It needs
to download the related css and js files accordingly to my local
website and then modify the link to the css and js files in the
downloaded website. This script is used to solve such a problem.